服务端提供登录功能

+++ common/src/main/java/com/cn/common/core/annotion/SocketModule.java
@@ -15,5 +15,5 @@ public @interface SocketModule {
      * 请求的模块号
      * @return
      */
-    short module();
+    int module();
 }
